Code review fixes: security, correctness, and cleanup
Security:
- safe_path: replace str.startswith with Path.is_relative_to — the old check
  was defeated by sibling directories whose names share a prefix with the base
- new_project: apply safe_path on the project name, matching load()

Correctness:
- export_digitakt: remove dead import that implied a path check that wasn't there
- tickUI: fix playhead off-by-one (prev default was 0 → highlighted step 1
  while audio played step 0; changed default to -1)
- togglePlay stop: reset _tickPos so playhead syncs with sequencer on restart
- toggleFill: await the POST and roll back UI state on failure

Cleanup:
- _trig_condition_passes: if chain → if/elif; fallthrough now returns False
  instead of silently accepting unknown condition strings
- sampler: read audio as float32 (matches audio engine dtype) to eliminate
  a float64→float32 conversion on every trig event inside the audio lock

+++ sila/engine/sampler.py
@@ -27,7 +27,7 @@ class LoadedSample:
         # Resample to target SR if needed.
         if sr != TARGET_SR:
             audio = soxr.resample(audio, sr, TARGET_SR, quality="HQ")
-        self.audio = audio  # float64 mono, target SR
+        self.audio = audio  # float32 mono, target SR
 
     def slice(self) -> np.ndarray:
         """Return the start–end region of the audio buffer."""
@@ -51,7 +51,7 @@ class SamplePlayer:
             src = safe_path(samples_dir, layer.path)
             if not src.exists():
                 continue
-            data, sr = sf.read(str(src), dtype="float64", always_2d=True)
+            data, sr = sf.read(str(src), dtype="float32", always_2d=True)
             mono = data[:, 0] if data.shape[1] == 1 else data.mean(axis=1)
             self._layers.append(LoadedSample(layer, mono, sr))
